backend/requirements.txt
# Python dependencies for Tic-Tac-Toe Web UI Backend

# Flask web framework
Flask>=3.0.0

# WebSocket support
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0

# CORS support
Flask-CORS>=4.0.0

# Raspberry Pi GPIO
RPi.GPIO>=0.7.1

# WSGI server for production
gevent>=23.9.0
gevent-websocket>=0.10.1
